+++ app/Domain/Stablecoin/Events/StablecoinMinted.php
@@ -12,7 +12,7 @@ class StablecoinMinted extends ShouldBeStored
         public readonly string $position_uuid,
         public readonly string $account_uuid,
         public readonly string $stablecoin_code,
-        public readonly int $amount,
+        public readonly float $amount,
         public readonly array $metadata = []
     ) {
     }
